feat: rename `search` query param to `filter`, adapt `ResourceDetailsControls` to also be used for dashboards
