+++ src/components/visualization/chart-table-filter/ChartTableFilter.tsx
@@ -22,8 +22,8 @@ export default function ChartTableFilter({
 }) {
   const searchParams = useSearchParams();
   const [filters, setFilters] = useState(
-    searchParams.get('search')
-      ? (JSON.parse(searchParams.get('search') ?? '{}') as Record<string, Filter>)
+    searchParams.get('filter')
+      ? (JSON.parse(searchParams.get('filter') ?? '{}') as Record<string, Filter>)
       : resource.defaultFilters ?? {},
   );
   const pathname = usePathname();
@@ -33,7 +33,7 @@ export default function ChartTableFilter({
     // use query parameters or default filter on initial render
     filter();
     // if query parameters are not set, set them to the value of defaultFilters
-    if (!searchParams.get('search') && resource.defaultFilters) {
+    if (!searchParams.get('filter') && resource.defaultFilters) {
       setParams(resource.defaultFilters);
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
@@ -64,9 +64,9 @@ export default function ChartTableFilter({
   const setParams = useDebouncedCallback((updatedFilters: Record<string, Filter>) => {
     const params = new URLSearchParams(searchParams);
     if (Object.keys(updatedFilters).length === 0) {
-      params.delete('search');
+      params.delete('filter');
     } else {
-      params.set('search', JSON.stringify(updatedFilters));
+      params.set('filter', JSON.stringify(updatedFilters));
     }
     router.replace(`${pathname}?${params.toString()}`);
   }, 300);
