+++ src/messages/en.json
@@ -23,6 +23,20 @@
         "allEntriesLabel": "Row",
         "contains": "contains"
     },
+    "DetailsPageControls": {
+        "download": "Download",
+        "embed": "Embed",
+        "fullscreen": "Fullscreen"
+    },
+    "EmbedModal": {
+        "dashboardBody": "This dashboard can be embedded using the following code.",
+        "dashboardTitle": "Embed dashboard",
+        "embedHeight": "Height",
+        "embedWidth": "Width",
+        "keepSearchParams": "Preserve search parameters",
+        "resourceBody": "The visualization for this resource can be embedded using the following code.",
+        "resourceTitle": "Embed resource visualization"
+    },
     "Error": {
         "configurationErrorMessage": "The application returned an error while trying to load and validate the configuration file(s).\n\nError: \"{error}\"",
         "configurationErrorTitle": "The application configuration is invalid or could not be loaded!",
@@ -68,16 +82,6 @@
     },
     "ResetView": {
         "resetTooltip": "Reset view"
-    },
-    "ResourceDetailsControls": {
-        "download": "Download",
-        "embed": "Embed",
-        "embedBody": "The visualization for this resource can be embedded using the following code.",
-        "embedHeight": "Height",
-        "embedTitle": "Embed resource visualization",
-        "embedWidth": "Width",
-        "fullscreen": "Fullscreen",
-        "keepSearchParams": "Preserve search parameters"
     },
     "Search": {
         "emptyLabel": "No entries found.",
